% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remisc.R
\name{acf}
\alias{acf}
\title{tidyverse-like \code{stats::acf} replacement}
\usage{
acf(x, max_lag = NULL, demean = TRUE)
}
\arguments{
\item{x}{Numeric vector}

\item{max_lag}{\code{stats::acf} default is \code{10 * log10(length(x))}}
}
\description{
tidyverse-like \code{stats::acf} replacement
}

