% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{c2f}
\alias{c2f}
\title{Convert celsius to farenheit}
\usage{
c2f(x, dec = 1)
}
\arguments{
\item{x}{an integer or numeric vector}

\item{dec}{an integer specifying the number of decimals of rounding}
}
\description{
Convert celsius to farenheit
}
\examples{
# Convert from celsius to fahrenheit
temp_c = c(0.00, 21.31, 100.00)
temp_f = c2f(temp_c, dec = 1)
}
