% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/string_manipulation.R
\name{capwords}
\alias{capwords}
\title{Convert mixed case strings to capital case}
\usage{
capwords(s, strict = FALSE)
}
\arguments{
\item{s}{A character vector of strings}

\item{strict}{Should strings such as acronyms also be converted}
}
\value{
A character vector with strings in mixed case converted to capital case \code{s}
}
\description{
Convert mixed case strings to capital case
}
\examples{
fish_names = c("king = chinook", "silver = coho")
capwords(fish_names)
}
