% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/short_generic.R
\name{check_max_char}
\alias{check_max_char}
\title{Check max length of strings in columns of a dataframe}
\usage{
check_max_char(dat, return_names = c("col_name", "max_chars"))
}
\arguments{
\item{dat}{a dataframe that can include a mix of datatypes}

\item{return_names}{A vector of two names for the return
dataframe.}
}
\value{
A dataframe reporting the \code{max(nchar(0)} of
  all columns in \code{data}. Non-character datatypes
  will return NA.
}
\description{
Determine the maximum length of each set of strings
  in the columns of a dataframe. Non-character values will be
  reported as NA.
}
\examples{
# Create example dataframe
fish_report = tibble::tibble(
  species = c("coho", "coho", "chin", "steelhead"),
  fish_method = c("Gillnet", "Seine", "Fish trap",
                  "Recreational angler"),
  total_length = c(660, 589, 815, 450))

# Report max length of strings in each column
fish_nchar = check_max_char(fish_report)

}
