% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remisc.R
\name{color_hue}
\alias{color_hue}
\title{Color helper for use with ggplot}
\usage{
color_hue(n, deg = 15, l = 65)
}
\arguments{
\item{n}{Number of colors}

\item{deg}{Degrees away from 0 on the color wheel}

\item{l}{Luminosity}
}
\description{
Color helper for use with ggplot
}

