% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remisc.R
\name{edf}
\alias{edf}
\title{Miscellaneous Helper Functions}
\usage{
edf(x, adjust = TRUE)
}
\arguments{
\item{x}{Numeric vector}

\item{adjust}{Logical.  Adjust so that \code{F(max(x)) < 1}?}
}
\description{
Calculate the empirical cumulative distribution function.
}
\examples{
edf(rnorm(9))
}

