% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_manipulation.R
\name{fish_stat_week}
\alias{fish_stat_week}
\title{Calculate statistical week from date values}
\usage{
fish_stat_week(dts, start_day = "Mon")
}
\arguments{
\item{dts}{A vector of date values formatted as character}

\item{start_day}{Starting day of the stat week. Options are "Mon" or "Sun"}
}
\description{
Generates a flavor of statistical week values typically encountered
  in fisheries management related applications. Allows the start date of the stat
  week to be set to either Sunday or Monday. Stat weeks will always be numbered
  starting at one on the first day of the year.
}
\examples{
# Generate a vector of date values
dtv = seq(as.Date("2017-01-01"),
          as.Date("2017-01-15"),
          by = "day")

# Generate vector of stat week values
stat_week = fish_stat_week(dtv)

# Create dataframe with date values
fish_catch = dplyr::data_frame(survey_date = dtv,
                               catch = as.integer(abs(rnorm(length(dtv)) * 10)))

# Add a stat_week column to fish_catch
fish_catch$stat_week = fish_stat_week(fish_catch$survey_date)

}
