% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/string_manipulation.R
\name{get_text_item}
\alias{get_text_item}
\title{Extract portion of a string defined be a separator}
\usage{
get_text_item(x, item = 2, sep = " ")
}
\arguments{
\item{x}{A vector of character strings}

\item{item}{The relative position of the item to extract}

\item{sep}{The separator used to split the string}
}
\description{
Extract portions of a string as defined by a separator such as a
  comma or whitespace. For example, in a vector strings of first and last
  name, such as "Bob Smith", the first name can be extracted as item number
  one and last name as item number two.
}
\examples{
# Define a vector of names that need to be broken out into columns
fish_names = tibble::tibble(names = c("Coho Salmon, Silvers",
                                      "Chinook Salmon, Kings"))

# Pull out the common name
fish_names$common_name = get_text_item(fish_names$names,
                                       item = 1, sep = " ")

# Pull out the alternate common name
fish_names$alternate_name = get_text_item(fish_names$names,
                                          item = 2, sep = ",")

# Trim whitespace from alternate_name
fish_names$alternate_name = trimws(fish_names$alternate_name, "b")
}
