% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remisc.R
\name{make_x}
\alias{make_x}
\title{User-friendly X-vector generator for DOE
Ironically, not factors}
\usage{
make_x(levels = c("lo", "hi"), reps = 1, n = length(levels) * reps)
}
\description{
User-friendly X-vector generator for DOE
Ironically, not factors
}

