% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{n_omit}
\alias{n_omit}
\title{Simple utility functions for vectors}
\usage{
n_omit(x)
}
\arguments{
\item{x}{a vector of any type}
}
\value{
The length of vector \code{x} ignoring an NA values.
}
\description{
These functions are intended for my convenience, and as an exercise to learn
package building. Functions take either a character vector in the case of
(trim, set_na, set_empty, trim_semi_colon) or a numeric or integer vector
in the case of f2c or c2f. The n_omit function will accept vectors of any
type
Get length of a vector, omitting NAs
}
\examples{
# Compute length of a vector, omitting NAs
fish_length = c(24.5, 16.8, 12.6, NA, 18.9)
n_omit(fish_length)
}
