% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remisc-package.R
\docType{package}
\name{remisc.package}
\alias{remisc.package}
\alias{remisc.package-package}
\title{remisc}
\description{
Package providing a variety of simple functions, primarily for personal use.
}
\author{
Are Strom

Maintainer: Are Strom <are.strom@gmail.com>
}
\keyword{package}
