% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{set_empty}
\alias{set_empty}
\title{Convert NAs to empty strings ("")}
\usage{
set_empty(x)
}
\arguments{
\item{x}{a character vector}
}
\description{
Convert NAs to empty strings ("")
}
\examples{
# Convert NA string values to empty ("") string.
fish_na = c("STHD", NA, "CHUM", "CHIN")
fish_missing = set_empty(fish_na)
}
