% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/short_generic.R
\name{set_na}
\alias{set_na}
\title{Convert empty strings, or indicators of missing values to NAs}
\usage{
set_na(x, na_value = "")
}
\arguments{
\item{x}{a character vector}

\item{na_value}{a string value such as "" or "9999" that should be converted
to NA}
}
\value{
A vector \code{x} of strings with indicators of missing values
  converted to NA
}
\description{
Convert empty strings, or indicators of missing values to NAs
}
\examples{
# To convert blanks to NAs in a vector
fish_missing = c("STHD", "", "CHUM", "CHIN")
fish_na = set_na(fish_missing)

# To convert all blanks in a dataframe to NAs
fish_barcode_out = FishBarcode
fish_barcode_out[] = lapply(fish_barcode_out, set_na)

}
