% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/short_generic.R
\name{set_na_zero}
\alias{set_na_zero}
\title{Set NA values in an integer or numeric vector to zero.}
\usage{
set_na_zero(x)
}
\arguments{
\item{x}{an integer or numeric vector}
}
\value{
A vector \code{x} of integers or numeric values equal to \code{0}.
 Use subsetting to apply function to only numeric or integer values.
}
\description{
Useful for converting NA values to zero when computing summary
 counts
}
\examples{
# Create example dataframe
fish_zero = dplyr::data_frame(fish_count = c(4L, NA, 5L),
                              water_temp = c(5.4, 6.2, NA))

# Set missing values to zero
fish_zero[] = lapply(fish_zero, set_na_zero)
}
