% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{trim}
\alias{trim}
\title{Trim whitespace before and after string}
\usage{
trim(x)
}
\arguments{
\item{x}{a character vector}
}
\value{
The length of vector \code{x} ignoring an NA values.
}
\description{
Trim whitespace before and after string
}
\examples{
# Trim whitespace before and after string value
fish_code = c("STHD ", " CHUM", " PINK ", "COHO")
trim(fish_code)
}
