% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_manipulation.R
\name{uuid_gen}
\alias{uuid_gen}
\title{Create a generator of Universally Unique Identifiers (UUIDs)}
\usage{
uuid_gen(more_state = "")
}
\arguments{
\item{more_state}{Additional items that can be added to generate random
state, such as Sys.info() (already used to generate state).}
}
\description{
Creates a generator of Universally Unique IDentifiers (UUIDs).
  Returns a parameterless function (with some randomish internal state),
  a call to which returns a vector of length one containing
  the character representation of a Version 4 UUID (RFC 4122).
}
\author{
Mikko Korpela
}
