% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/short_generic.R
\name{c2f}
\alias{c2f}
\title{Convert temperature in celsius to degrees farenheit}
\usage{
c2f(x, dec = 1)
}
\arguments{
\item{x}{an integer or numeric vector of temperature measurements in degrees celsius}

\item{dec}{an integer specifying the number of decimals of rounding}
}
\value{
A vector \code{x} of temperature measurements converted from
  celsius to fahrenheit. The number of decimals returned is set using
  \code{dec}.
}
\description{
Convert temperature in celsius to degrees farenheit
}
\examples{
temp_c = c(0.00, 21.31, 100.00)
temp_f = c2f(temp_c, dec = 1)
}
