% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{f2c}
\alias{f2c}
\title{Convert farenheit to celsius}
\usage{
f2c(x, dec = 1)
}
\arguments{
\item{x}{an integer or numeric vector}

\item{dec}{an integer specifying the number of decimals of rounding}
}
\value{
A vector \code{x} of temperature measurements converted from
  fahrenheit to celsius. The number of decimals returned is set using
  `dec`.
}
\description{
Convert farenheit to celsius
}
\examples{
temp_f = c(32, 70.356, 212)
temp_c = f2c(temp_f, dec = 2)
}
