% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/short_generic.R
\name{mult_by}
\alias{mult_by}
\title{Multiply vectors in a dataframe by a numeric or integer value}
\usage{
mult_by(dat, mult_value)
}
\arguments{
\item{dat}{a dataframe with only integer or numeric vectors}

\item{mult_value}{A numeric or integer value to use as a multiplier}
}
\value{
A dataframe \code{dat} with specified columns multiplied by
  \code{mult_value}.
}
\description{
Multiply vectors in a dataframe by a numeric or integer value
}
\examples{
fish_lengths = tibble::tibble(species = c("coho", "coho", "chin", "pink"),
                              fork_length = c(650, 580, 804, NA),
                              total_length = c(660, 589, 815, 450))
fish_lengths[,2:3] = mult_by(fish_lengths[,2:3], 0.1)
}
