% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/short_generic.R
\name{set_empty}
\alias{set_empty}
\title{Convert NAs to empty strings ("")}
\usage{
set_empty(x)
}
\arguments{
\item{x}{a character vector}
}
\value{
A vector \code{x} of strings with NAs converted to an empty string
  ("")
}
\description{
This can be useful when exporting a dataframe
  to a csv or excel format. It gets rid of the printed NA values in the output.
}
\examples{
# Set NA values in a vector to empty strings
fish_na = c("STHD", NA, "CHUM", "CHIN")
fish_missing = set_empty(fish_na)

# Set NA values in an entire dataframe to empty strings
fish_na = dplyr::data_frame(fish_day = c("Mon", NA_character_, "Wed", "Thur"),
                            fish_sp = c("STHD", NA_character_, "CHUM", "CHIN"))
fish_na[] = lapply(fish_na, set_empty)
}
