% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/short_generic.R
\name{set_empty}
\alias{set_empty}
\title{Convert NAs to empty strings ("")}
\usage{
set_empty(x)
}
\arguments{
\item{x}{a character vector}
}
\value{
A vector \code{x} of strings with NAs converted to an empty string
  ("")
}
\description{
Convert NAs to empty strings ("")
}
\examples{
fish_na = c("STHD", NA, "CHUM", "CHIN")
fish_missing = set_empty(fish_na)
}
