% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/short_generic.R
\name{trim}
\alias{trim}
\title{Trim white-space before and after each string in a vector}
\usage{
trim(x)
}
\arguments{
\item{x}{a character vector}
}
\value{
A vector \code{x} of strings with white-space removed
}
\description{
Trim white-space before and after each string in a vector
}
\examples{
fish_code = c("STHD ", " CHUM", " PINK ", "COHO")
trim(fish_code)
}
