% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{trim_semi_colon}
\alias{trim_semi_colon}
\title{Trim trailing semi-colon from a vector of strings}
\usage{
trim_semi_colon(x)
}
\arguments{
\item{x}{a character vector}
}
\description{
Trim trailing semi-colon from a vector of strings
}
\examples{
# Get rid of a trailing semi-colon
observers = c("Austin;Wesley;", "Losee;Madel", "Morris;Starling;")
observers_fixed = trim_semi_colon(observers)
}
