% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/short_generic.R
\name{trim_semi_colon}
\alias{trim_semi_colon}
\title{Trim trailing semi-colon from each string in a vector of strings}
\usage{
trim_semi_colon(x)
}
\arguments{
\item{x}{a character vector}
}
\value{
A vector \code{x} of strings with trailing semi-colons stripped out
}
\description{
Trim trailing semi-colon from each string in a vector of strings
}
\examples{
observers = c("Austin;Wesley;", "Losee;Madel", "Morris;Starling;")
observers_fixed = trim_semi_colon(observers)
}
